requirements.txt
Django==5.2
djangorestframework==3.14.0
dnspython==2.4.2
python-whois==0.8.0
requests==2.25.1
beautifulsoup4==4.9.3
python-dotenv==1.0.0
openai==1.12.0
reportlab==4.1.0
fpdf2==2.7.8
googlesearch-python==1.1.0
httpx==0.27.2
python-nmap==0.7.1